requirements.txt
python-dotenv
pandas
numpy==1.24.4
pymongo
scikit-learn==1.2.2
scipy==1.10.1
certifi
dill
pyyaml
xgboost

#-e . 